Add gRPC server lib
